move: operatorフォルダを作成
BinaryOperatorを継承するクラスを別フォルダに移動
