+++ cli.py
@@ -2,7 +2,7 @@ from calculator.calculator import Calculator
 
 END_CMD = ['q', 'quit', 'exit']
 
-if __name__ == '__main__':
+def calculator_cli():
     calc = Calculator()
 
     while(True):
@@ -10,8 +10,11 @@ if __name__ == '__main__':
         s = input()
 
         if s in END_CMD:
-            exit
+            return
 
         # TODO: インプットが正しいかのチェックを別クラスにする
         print(calc.process(s.split()))
+
+if __name__ == '__main__':
+    calculator_cli()
